+++ backend/app/configuration/domains/model.py
@@ -1,15 +1,32 @@
 import uuid
 from typing import TYPE_CHECKING
 
-from sqlalchemy import Column, String
+from sqlalchemy import (
+    Column,
+    DateTime,
+    ForeignKey,
+    String,
+    Table,
+)
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, Session, relationship
 
 from app.database.database import Base, ensure_exists
-from app.shared.model import BaseORM
+from app.shared.model import BaseORM, utcnow
 
 if TYPE_CHECKING:
     from app.abstract_data_product.model import AbstractDataProduct
+    from app.configuration.environments.model import Environment
+
+
+domain_environment_table = Table(
+    "domain_environments",
+    Base.metadata,
+    Column("domain_id", ForeignKey("domains.id")),
+    Column("environment_id", ForeignKey("environments.id")),
+    Column("created_on", DateTime(timezone=False), server_default=utcnow()),
+    Column("updated_on", DateTime(timezone=False), onupdate=utcnow()),
+)
 
 
 class Domain(Base, BaseORM):
@@ -21,6 +38,9 @@ class Domain(Base, BaseORM):
 
     abstract_data_products: Mapped[list["AbstractDataProduct"]] = relationship(
         "AbstractDataProduct", lazy="raise"
+    )
+    environments: Mapped[list["Environment"]] = relationship(
+        secondary=domain_environment_table, lazy="raise"
     )