+++ cli/golang/pkg/api/oas_request_encoders_gen.go
@@ -749,6 +749,20 @@ func encodeUpdateDomainRequest(
 	return nil
 }
 
+func encodeUpdateEnvironmentIsGlobalRequest(
+	req *EnvironmentUpdateGlobal,
+	r *http.Request,
+) error {
+	const contentType = "application/json"
+	e := new(jx.Encoder)
+	{
+		req.Encode(e)
+	}
+	encoded := e.Bytes()
+	ht.SetBody(r, bytes.NewReader(encoded), contentType)
+	return nil
+}
+
 func encodeUpdateOutputPortRequest(
 	req *DatasetUpdate,
 	r *http.Request,